Add support for running API test with helm
This change introduces support for running Harbor API tests by using the
helm test command.

The tests are executed in a pod with 2 containers, one for the test
execution itself and the other one running docker daemon. The tests are
the same API tests executed by the Harbor upstream project.

Note: to allow anyone to test this, change the images used for harbor
tests are referencing home/flaviosr. Before merging the test images
needs to be updated with the correct values in the `values.yaml` file.