Remove trivy helm chart changes from the 2.0 baseline
The 1.4.0 helm chart version in the release-2.0 branch could be used
as a baseline for deploying the 2.0.0 SUSE Registry Tech Preview
version using the images in Devel:CAPS:Registry. However, some
trivy related changes crept through and need to be removed (trivy
is not available in the Tech Preview release).