Tests: Set NOTARY_URL when using notary with ingress
By default the tests expect notary to be served on the 4443 port,
however when using an ingress it is served on 443. This change sets the
`NOTARY_URL` environment variable on the api-tests container with the
value defined on expose.ingress.hosts.notary so the tests are executed
using the correct notary url.