Rework CI workflows
The following changes were made:
* **New workflow `[Trigger] OBS Build`**: this workflow consists of one job
which is responsible for triggering the Registry CI workflow whenever
there is a new build of the containers or charts repositories on the
specified OBS projects. This jobs runs on a Leap 15.2 container and it
should be always running. The trigger happens by creating a
`workflow_dispatch` event using the github API (see [1]).
* **The Registry CI workflows are now triggered by OBS repository
builds**: With the addition of the `[Trigger] OBS Build` workflow there
is no need to execute the Registry CI workflows on a schedule anymore.
* **The Registry CI jobs are now executed inside a container**:
this enable better isolation between jobs ensuring that a fresh and
dedicated environment is available where requirements can be installed
on demand. It also removes the need to install specific requirements on
the ci workers and prevent version conflict between different jobs
requirements.
* **GitHub deployments were removed**: there is no use for it.
* **Notification to RocketChat**: the Registry CI workflows now reports
the job status (and other details about it) by sending a notification
to the `#ecosystem-ci` channel on rocketchat.
* **Renamed `[CaaSP 4] Rotate Cluster` workflow to `Rotate Kubernetes cluster`**:
the goal is to use the same workflow to include jobs deploying other
kubernetes clusters (not only CaaSP) in the future.
* **Added a logging stack to the CaaSP cluster**: the `deploy_caasp` job
now also deploys a logging stack composed of Elasticsearch, Fluent-bit
and Kibana. For that a new directory `cluster-services` was created
under the `.github` directory to store services that can be deployed
into the kubernetes cluster.
* Fix other minor issues regarding the kubernetes cluster rotation
logic.

1. https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#workflow_dispatch